refactor: Redefinde folder structure
