+++ src/common/BookInterface.js
