+++ src/home/BookShelf.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Book from './Book';
+import Book from '../common/Book';
 
 const BookShelf = props => {
   const { activeShelf, bookShelves, books } = props;
